rename: 검색 컴포넌트들 폴더 이동
- 폴더 이동: common -> domain
- 파일명 변경: KeywordsContext -> SearchContext
