+++ src/components/domain/Search/index.tsx
@@ -1,7 +1,7 @@
-import SearchList from '../common/Search/SearchList';
-import SearchBar from '../common/Search/SearchBar';
-import { useKeywordsContext } from './KeywordsContext';
-import { useStationInfo } from '../../api/stations';
+import SearchList from './SearchList';
+import SearchBar from './SearchBar';
+import { useKeywordsContext } from './SearchContext';
+import { useStationInfo } from '../../../api/stations';
 
 const Search = () => {
   const { keywords } = useKeywordsContext();
@@ -13,7 +13,7 @@ const Search = () => {
         placeholder='역이름을 입력해주세요.'
         listeningMessage='듣고 있습니다! 역이름을 말해주세요.'
       />
-      <SearchList label='최근 검색' data={data} />
+      <SearchList data={data} />
     </>
   );
 };
