--- src/pages/Search.tsx
@@ -1,12 +1,0 @@
-import { KeywordsContextProvider } from '../components/domain/KeywordsContext';
-import Search from '../components/domain/Search';
-
-const SearchPage = () => {
-  return (
-    <KeywordsContextProvider>
-      <Search />
-    </KeywordsContextProvider>
-  );
-};
-
-export default SearchPage;
